+++ stac_fastapi/pgstac/config.py
@@ -202,6 +202,7 @@ class Settings(ApiSettings):
     enabled_extensions: str = ""
     enable_transactions_extensions: bool = False
     enable_catalogs_extension: bool = False
+    hide_alternate_parents: bool = False
     validate_extensions: bool = False
     """
     Validate `stac_extensions` schemas against submitted data when creating or updated STAC objects.